feat(Board, BoardInitializer): 저장된 게임을 확인하고 여부를 받는 게임 로직 추가 및 BoardInitializer를 인터페이스로 변경
