+++ src/test/java/chess/domain/board/BoardTest.java
@@ -17,7 +17,7 @@ class BoardTest {
     @Test
     @DisplayName("보드에서 체스 말을 이동시킬 수 있다.")
     void movePiece() {
-        Board board = new Board(new BoardInitializer());
+        Board board = new Board(new ClearBoardInitializer());
         board.tryMove(Position.of(1, 7), Position.of(1, 5));
         Map<Position, Piece> boardPieces = board.getBoard();
         assertAll(
@@ -30,7 +30,7 @@ class BoardTest {
     @Test
     @DisplayName("보드에서 체스 말을 이동할 수 없는 경우 예외가 발생한다.")
     void movePieceThrowException() {
-        Board board = new Board(new BoardInitializer());
+        Board board = new Board(new ClearBoardInitializer());
         assertThatThrownBy(() -> board.tryMove(Position.of(1, 7), Position.of(1, 2)))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("이동이 불가능한 위치입니다.");
@@ -40,7 +40,7 @@ class BoardTest {
     @Test
     @DisplayName("움직이려는 말의 진영 차례가 아닌 경우 예외가 발생한다.")
     void isTurnThrowException() {
-        Board board = new Board(new BoardInitializer());
+        Board board = new Board(new ClearBoardInitializer());
         assertThatThrownBy(() -> board.tryMove(Position.of(1, 2), Position.of(1, 4)))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("해당 진영의 차례가 아닙니다.");
@@ -50,14 +50,14 @@ class BoardTest {
     @Test
     @DisplayName("King이 잡힐 시에 게임이 종료된다.")
     void isNotFinish() {
-        Board board = new Board(new BoardInitializer());
+        Board board = new Board(new ClearBoardInitializer());
         assertThat(board.isFinish()).isFalse();
     }
 
     @Test
     @DisplayName("King이 잡힐 시에 게임이 종료된다.")
     void isFinish() {
-        Board board = new Board(new BoardInitializer());
+        Board board = new Board(new ClearBoardInitializer());
 
         // RNBQqB.R <- king이 죽는 기보
         // PPPP..PP
